+++ engine/renderer/tr_arb.cpp
@@ -1769,7 +1769,7 @@ static void R_Bloom_LensEffect( float alpha )
 	vec4_t color;
 
 	alpha /= (float)ARRAY_LEN( lc );
-	for ( i = 0; i < ARRAY_LEN( lc ); i++ ) {
+	for ( i = 0; (size_t)i < ARRAY_LEN( lc ); i++ ) {
 		VectorCopy( lc[i], color ); color[3] = alpha;
 		R_Setup_Quad_Lens( (i+1)*144, color, &verts[i*6], &coords[i*6], &colors[i*6] );
 	}
